+++ com.oracle.truffle.r.nodes.builtin/src/com/oracle/truffle/r/nodes/builtin/base/Parse.java
@@ -30,14 +30,7 @@ import static com.oracle.truffle.r.runtime.builtins.RBuiltinKind.INTERNAL;
 import java.io.File;
 import java.io.IOException;
 import java.net.URISyntaxException;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.EnumSet;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Map;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 
 import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
 import com.oracle.truffle.api.TruffleFile;
@@ -46,7 +39,6 @@ import com.oracle.truffle.api.source.Source;
 import com.oracle.truffle.api.source.SourceSection;
 import com.oracle.truffle.r.nodes.attributes.SetFixedAttributeNode;
 import com.oracle.truffle.r.nodes.builtin.RBuiltinNode;
-import com.oracle.truffle.r.runtime.RDeparse;
 import com.oracle.truffle.r.runtime.RError;
 import com.oracle.truffle.r.runtime.RInternalError;
 import com.oracle.truffle.r.runtime.RRuntime;
@@ -59,24 +51,18 @@ import com.oracle.truffle.r.runtime.conn.RConnection;
 import com.oracle.truffle.r.runtime.conn.RConnection.ReadLineWarning;
 import com.oracle.truffle.r.runtime.conn.StdConnections;
 import com.oracle.truffle.r.runtime.context.Engine.ParseException;
+import com.oracle.truffle.r.runtime.context.Engine.ParsedExpression;
+import com.oracle.truffle.r.runtime.context.Engine.ParserMetadata;
 import com.oracle.truffle.r.runtime.context.RContext;
 import com.oracle.truffle.r.runtime.data.RComplex;
 import com.oracle.truffle.r.runtime.data.RDataFactory;
-import com.oracle.truffle.r.runtime.data.REmpty;
 import com.oracle.truffle.r.runtime.data.RExpression;
 import com.oracle.truffle.r.runtime.data.RIntVector;
-import com.oracle.truffle.r.runtime.data.RMissing;
 import com.oracle.truffle.r.runtime.data.RNull;
 import com.oracle.truffle.r.runtime.data.RPairList;
 import com.oracle.truffle.r.runtime.data.RSymbol;
 import com.oracle.truffle.r.runtime.data.model.RAbstractStringVector;
 import com.oracle.truffle.r.runtime.env.REnvironment;
-import com.oracle.truffle.r.runtime.nodes.RSyntaxCall;
-import com.oracle.truffle.r.runtime.nodes.RSyntaxConstant;
-import com.oracle.truffle.r.runtime.nodes.RSyntaxElement;
-import com.oracle.truffle.r.runtime.nodes.RSyntaxFunction;
-import com.oracle.truffle.r.runtime.nodes.RSyntaxLookup;
-import com.oracle.truffle.r.runtime.nodes.RSyntaxVisitor;
 
 /**
  * Internal component of the {@code parse} base package function.
@@ -152,7 +138,8 @@ public abstract class Parse extends RBuiltinNode.Arg6 {
         }
         try {
             Source source = srcFile != RNull.instance ? createSource(srcFile, coalescedLines) : createSource(conn, coalescedLines);
-            RExpression exprs = RContext.getEngine().parse(source);
+            ParsedExpression parseRes = RContext.getEngine().parse(source);
+            RExpression exprs = parseRes.getExpression();
             if (n > 0 && n < exprs.getLength()) {
                 Object[] subListData = new Object[n];
                 for (int i = 0; i < n; i++) {
@@ -162,7 +149,7 @@ public abstract class Parse extends RBuiltinNode.Arg6 {
             }
             // Handle the required R attributes
             if (srcFile instanceof REnvironment) {
-                addAttributes(exprs, source, (REnvironment) srcFile);
+                addAttributes(parseRes, exprs, source, (REnvironment) srcFile);
             }
             return exprs;
         } catch (ParseException ex) {
@@ -174,7 +161,9 @@ public abstract class Parse extends RBuiltinNode.Arg6 {
         StringBuilder sb = new StringBuilder();
         for (int i = 0; i < lines.getLength(); i++) {
             sb.append(lines.getDataAt(i));
-            sb.append('\n');
+            if (i < lines.getLength() - 1) {
+                sb.append('\n');
+            }
         }
         return sb.toString();
     }
@@ -246,7 +235,7 @@ public abstract class Parse extends RBuiltinNode.Arg6 {
         }
     }
 
-    private void addAttributes(RExpression exprs, Source source, REnvironment srcFile) {
+    private void addAttributes(ParsedExpression parseRes, RExpression exprs, Source source, REnvironment srcFile) {
         Object[] srcrefData = new Object[exprs.getLength()];
         for (int i = 0; i < srcrefData.length; i++) {
             Object data = exprs.getDataAt(i);
@@ -276,309 +265,12 @@ public abstract class Parse extends RBuiltinNode.Arg6 {
         setWholeSrcRefAttrNode.setAttr(exprs, RDataFactory.createIntVector(wholeSrcrefData, RDataFactory.COMPLETE_VECTOR));
         setSrcFileAttrNode.setAttr(exprs, srcFile);
 
-        RIntVector parseData = new ParseDataVisitor(exprs).getParseData();
+        ParserMetadata metadata = parseRes.getParseData();
+        RIntVector parseData = RDataFactory.createIntVector(metadata.getData(), false);
+        parseData.setAttr("tokens", RDataFactory.createStringVector(metadata.getTokens(), false));
+        parseData.setAttr("text", RDataFactory.createStringVector(metadata.getText(), false));
+        parseData.setClassAttr(RDataFactory.createStringVector("parseData"));
+        parseData.setDimensions(new int[]{8, parseData.getLength() / 8});
         srcFile.safePut("parseData", parseData);
-    }
-
-    /**
-     * This class aspires to reconstruct the original parse tree through visiting a parsed
-     * expression(s). The reconstruction cannot be done completely as some information is lost, but
-     * the result should suffice for templating packages, such as knitr.
-     */
-    static class ParseDataVisitor extends RSyntaxVisitor<ParseDataVisitor.OctetNode> {
-
-        static class OctetNode {
-            static final List<OctetNode> NO_CHILDREN = java.util.Collections.emptyList();
-
-            final int id;
-            final int startLine;
-            final int startColumn;
-            final int endLine;
-            final int endColumn;
-            final TokenType tokenType;
-            final String txt;
-            final List<OctetNode> children;
-
-            OctetNode(int id, int startLine, int startColumn, int endLine, int endColumn, TokenType tokenType, String txt, List<OctetNode> children) {
-                this.id = id;
-                this.startLine = startLine;
-                this.startColumn = startColumn;
-                this.endLine = endLine;
-                this.endColumn = endColumn;
-                this.tokenType = tokenType;
-                this.txt = txt;
-                this.children = children;
-            }
-
-            void store(ParseDataVisitor vis, int parentId) {
-                for (OctetNode child : children) {
-                    child.store(vis, id);
-                }
-
-                vis.data.add(startLine);
-                vis.data.add(startColumn);
-                vis.data.add(endLine);
-                vis.data.add(endColumn);
-                vis.data.add(tokenType.terminal ? 1 : 0);
-                vis.data.add(tokenType.code);
-                vis.data.add(id);
-                vis.data.add(parentId);
-
-                vis.tokens.add(tokenType.tokenName);
-                vis.text.add(txt);
-            }
-
-        }
-
-        private final RExpression exprs;
-
-        /**
-         * A list of parse data octets. Every octet corresponds to one term/non-term. The octet is
-         * composed as follows:
-         *
-         * <pre>
-         * line1 col1 line2 col2 terminal token id parent
-         * </pre>
-         *
-         */
-        private final List<Integer> data = new ArrayList<>();
-        private boolean containsNA = false;
-
-        private final List<String> tokens = new ArrayList<>();
-
-        private final List<String> text = new ArrayList<>();
-
-        private int idCounter = 0;
-
-        ParseDataVisitor(RExpression exprs) {
-            this.exprs = exprs;
-        }
-
-        /**
-         * This enum mimics the <code>yytokentype</code> enum from <code>src/main/gram.c</code>.
-         */
-        enum TokenType {
-            expr(77, false),
-            IF(272, true, "if"),
-            WHILE(274, true, "while"),
-            FOR(270, true, "for"),
-            IN(271, true, "in"),
-            BREAK(276, true, "break"),
-            REPEAT(277, true, "repeat"),
-            NEXT(275, true, "next"),
-            LEFT_ASSIGN(266, true, "<-"),
-            EQ_ASSIGN(267, true, "="),
-            RIGHT_ASSIGN(268, true, "->"),
-            PLUS(43, true, "+", true, true),
-            MINUS(45, true, "-", true, true),
-            MULT(42, true, "*", true, true),
-            DIV(47, true, "/", true, true),
-            POW(94, true, "^", true, true),
-            DOLLAR(36, true, "$", true, true),
-            AT(64, true, "@", true, true),
-            COLON(58, true, ":", true, true),
-            QUESTION(63, true, "?", true, false),
-            TILDE(126, true, "~", true, false),
-            EXCLAMATION(33, true, "!", true, false),
-            GT(278, true, ">"),
-            GE(279, true, ">="),
-            LT(280, true, "<"),
-            LE(281, true, "<="),
-            EQ(282, true, "=="),
-            NE(283, true, "!="),
-            AND(284, true, "&"),
-            AND2(286, true, "&&"),
-            OR(285, true, "|"),
-            OR2(287, true, "||"),
-            SYMBOL(263, true),
-            SYMBOL_FUNCTION_CALL(296, true),
-            SPECIAL(304, true),
-            NULL_CONST(262, true),
-            NUM_CONST(261, true),
-            STR_CONST(260, true),
-            FUNCTION(264, true),
-            LPAREN(40, true, "(", true, false),
-            RPAREN(41, true, ")", true, false);
-
-            final int code;
-            final boolean terminal;
-            final String symbol;
-            final String tokenName;
-            final boolean infix;
-
-            TokenType(int c, boolean term, String symbol, boolean useSymbolAsTokenName, boolean infix) {
-                this.code = c;
-                this.terminal = term;
-                this.symbol = symbol;
-                this.tokenName = useSymbolAsTokenName ? "'" + symbol + "'" : name();
-                this.infix = infix;
-            }
-
-            TokenType(int c, boolean term, String symbol) {
-                this(c, term, symbol, false, false);
-            }
-
-            TokenType(int c, boolean term) {
-                this(c, term, null, false, false);
-            }
-
-            String getSymbol() {
-                return symbol;
-            }
-        }
-
-        private static final Map<String, TokenType> FUNCTION_TOKENS = Arrays.stream(TokenType.values()).filter(tt -> tt.symbol != null).collect(
-                        Collectors.toMap(TokenType::getSymbol, Function.identity()));
-
-        private OctetNode newOctet(RSyntaxElement element, TokenType tokenType, String txt, List<OctetNode> children) {
-            return newOctet(element.getSourceSection().getStartLine(), element.getSourceSection().getStartColumn(), element.getSourceSection().getEndLine(), element.getSourceSection().getEndColumn(),
-                            tokenType, txt, children);
-        }
-
-        private OctetNode newOctet(RSymbol symbol, List<OctetNode> children) {
-            OctetNode octet = newOctet(RRuntime.INT_NA, RRuntime.INT_NA, RRuntime.INT_NA, RRuntime.INT_NA, TokenType.SYMBOL, symbol.getName(), children);
-            containsNA = true;
-            return octet;
-        }
-
-        private OctetNode newOctet(int startLine, int startColumn, int endLine, int endColumn, TokenType tokenType, String txt, List<OctetNode> children) {
-            return new OctetNode(++idCounter, startLine, startColumn, endLine, endColumn, tokenType, txt, children);
-        }
-
-        @TruffleBoundary
-        RIntVector getParseData() {
-            int exprLen = exprs.getLength();
-            List<OctetNode> rootOctets = new ArrayList<>();
-            for (int i = 0; i < exprLen; i++) {
-                Object x = exprs.getDataAt(i);
-                if ((x instanceof RPairList && ((RPairList) x).isLanguage())) {
-                    RSyntaxElement rep = ((RPairList) x).getSyntaxElement();
-                    rootOctets.add(accept(rep));
-                } else if (x instanceof RSymbol) {
-                    rootOctets.add(newOctet((RSymbol) x, OctetNode.NO_CHILDREN));
-                } else {
-                    // TODO: primitives
-                }
-            }
-
-            // Store the octet tree to the corresponding vectors
-            for (OctetNode rootOctet : rootOctets) {
-                rootOctet.store(this, 0);
-            }
-
-            int[] dataArray = new int[data.size()];
-            for (int i = 0; i < dataArray.length; i++) {
-                dataArray[i] = data.get(i);
-            }
-            RIntVector parseData = RDataFactory.createIntVector(dataArray, !containsNA);
-
-            String[] textArray = new String[text.size()];
-            for (int i = 0; i < textArray.length; i++) {
-                textArray[i] = text.get(i);
-            }
-
-            String[] tokensArray = new String[tokens.size()];
-            for (int i = 0; i < tokensArray.length; i++) {
-                tokensArray[i] = tokens.get(i);
-            }
-
-            parseData.setAttr("tokens", RDataFactory.createStringVector(tokensArray, true));
-            parseData.setAttr("text", RDataFactory.createStringVector(textArray, true));
-            parseData.setClassAttr(RDataFactory.createStringVector("parseData"));
-            parseData.setDimensions(new int[]{8, idCounter});
-
-            return parseData;
-        }
-
-        @Override
-        protected OctetNode visit(RSyntaxCall element) {
-            LinkedList<OctetNode> children = new LinkedList<>();
-
-            RSyntaxElement lhs = element.getSyntaxLHS();
-            if (lhs instanceof RSyntaxCall) {
-                children.add(accept(lhs));
-                children.addAll(visitArguments(element, 0, Integer.MAX_VALUE));
-            } else if (lhs instanceof RSyntaxLookup) {
-                String symbol = ((RSyntaxLookup) lhs).getIdentifier();
-                RDeparse.Func func = RDeparse.getFunc(symbol);
-                TokenType tt = null;
-                if (func != null) {
-                    tt = FUNCTION_TOKENS.get(symbol);
-                }
-                if (tt == null) {
-                    tt = TokenType.SYMBOL_FUNCTION_CALL;
-                }
-                if (tt.infix) {
-                    children.addAll(visitArguments(element, 0, 1));
-                    children.add(newOctet(lhs, tt, symbol, OctetNode.NO_CHILDREN));
-                    children.addAll(visitArguments(element, 1, 2));
-                } else {
-                    children.add(newOctet(lhs, tt, symbol, OctetNode.NO_CHILDREN));
-                    children.addAll(visitArguments(element, 0, Integer.MAX_VALUE));
-                }
-                if (tt == TokenType.LPAREN) {
-                    OctetNode lastChild = children.getLast();
-                    children.add(new OctetNode(++idCounter, lastChild.endLine, lastChild.endColumn + 1, lastChild.endLine, lastChild.endColumn + 1, TokenType.RPAREN, ")", OctetNode.NO_CHILDREN));
-                }
-            }
-
-            return newOctet(element, TokenType.expr, "", children);
-        }
-
-        private List<OctetNode> visitArguments(RSyntaxCall element, int from, int to) {
-            List<OctetNode> children = new ArrayList<>();
-            RSyntaxElement[] args = element.getSyntaxArguments();
-            for (int i = from; i < Math.min(args.length, to); i++) {
-                if (args[i] != null) {
-                    OctetNode argOctet = accept(args[i]);
-                    if (argOctet != null) {
-                        children.add(argOctet);
-                    }
-                }
-            }
-            return children;
-        }
-
-        @Override
-        protected OctetNode visit(RSyntaxConstant element) {
-            TokenType tt;
-            Object value = element.getValue();
-            if (value == RNull.instance) {
-                tt = TokenType.NULL_CONST;
-            } else if (value instanceof Number) {
-                tt = TokenType.NUM_CONST;
-            } else if (value instanceof String) {
-                tt = TokenType.STR_CONST;
-            } else if (value instanceof RComplex) {
-                tt = TokenType.NUM_CONST;
-            } else if (value == REmpty.instance || value == RMissing.instance || value == RSymbol.MISSING) {
-                return null;    // ignored
-            } else {
-                throw RInternalError.shouldNotReachHere("Unknown RSyntaxConstant in ParseDataVisitor " + (value == null ? "null" : value.getClass().getSimpleName()));
-            }
-
-            OctetNode constChild = newOctet(element, tt, element.getSourceSection().getCharacters().toString(), OctetNode.NO_CHILDREN);
-            return newOctet(element, TokenType.expr, "", java.util.Collections.singletonList(constChild));
-        }
-
-        @Override
-        protected OctetNode visit(RSyntaxLookup element) {
-            String symbol = element.getIdentifier();
-            return newOctet(element, TokenType.SYMBOL, symbol, OctetNode.NO_CHILDREN);
-        }
-
-        @Override
-        protected OctetNode visit(RSyntaxFunction element) {
-            List<OctetNode> children = new ArrayList<>();
-            for (RSyntaxElement arg : element.getSyntaxArgumentDefaults()) {
-                if (arg != null) {
-                    children.add(accept(arg));
-                }
-            }
-
-            children.add(accept(element.getSyntaxBody()));
-            return newOctet(element, TokenType.FUNCTION, "function", children);
-        }
-
     }
 }